add logs to rollback cli and some refactor
