+++ cot-cli/Cargo.toml
@@ -20,7 +20,6 @@ path = "src/main.rs"
 workspace = true
 
 [dependencies]
-anstyle.workspace = true
 anyhow.workspace = true
 cargo_toml.workspace = true
 chrono.workspace = true
