+++ cot/Cargo.toml
@@ -17,6 +17,7 @@ workspace = true
 
 [dependencies]
 aide = { workspace = true, optional = true }
+anstyle.workspace = true
 askama = { workspace = true, features = ["std"] }
 async-trait.workspace = true
 axum = { workspace = true, features = ["http1", "tokio"] }
