+++ cot/src/utils.rs
@@ -1,4 +1,6 @@
+//! Cot utilities
 pub(crate) mod accept_header_parser;
 pub(crate) mod chrono;
+pub mod cli;
 #[cfg(feature = "db")]
 pub(crate) mod graph;
